+++ go.mod
@@ -1,12 +1,16 @@
 module clammit
 
-go 1.15
+go 1.23.2
+
+require (
+	github.com/dutchcoders/go-clamd v0.0.0-20170520113014-b970184f4d9e
+	github.com/stretchr/testify v1.10.0
+	gopkg.in/gcfg.v1 v1.2.3
+)
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/dutchcoders/go-clamd v0.0.0-20170520113014-b970184f4d9e
-	github.com/go-delve/delve v1.6.0 // indirect
-	github.com/stretchr/testify v1.7.0
-	gopkg.in/gcfg.v1 v1.2.3
+	github.com/pmezard/go-difflib v1.0.0 // indirect
 	gopkg.in/warnings.v0 v0.1.2 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
